+++ valid/PositiveMap.scala
@@ -3,7 +3,7 @@ import leon.lang._
 object PositiveMap {
   
   abstract class List
-  case class Cons(head: Int, tail: List) extends List
+  case class Cons(head: BigInt, tail: List) extends List
   case class Nil() extends List
 
   def positive(list: List): Boolean = list match {
@@ -11,7 +11,7 @@ object PositiveMap {
     case Nil() => true
   }
 
-  def positiveMap_passing_1(f: (Int) => Int, list: List): List = {
+  def positiveMap_passing_1(f: (BigInt) => BigInt, list: List): List = {
     list match {
       case Cons(head, tail) =>
         val fh = f(head)
@@ -21,7 +21,7 @@ object PositiveMap {
     }
   } ensuring { res => positive(res) }
 
-  def positiveMap_passing_2(f: (Int) => Int, list: List): List = {
+  def positiveMap_passing_2(f: (BigInt) => BigInt, list: List): List = {
     list match {
       case Cons(head, tail) =>
         val fh = f(head)
